Add files via upload
The Tic Tac Toe game is a widely recognized and simple two-player game that provides an excellent opportunity to implement graphical user interfaces and core game logic in Java. This project aims to recreate the Tic Tac Toe experience using Java Swing, a graphical framework for Java that allows the creation of rich user interfaces.